move Example1 from main Sql file
